requirements.txt
asyncpg==0.27.0
fastapi==0.92.0
pydantic==1.10.12
redis==5.0.0
python-dotenv==1.0.0